requirements.txt
discord.py
python-dotenv
aiohttp
wavelink
pynacl
yt-dlp